+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Fixed
     - iRODS file search fails with upper case search terms (#2560)
     - Vue3app initial routing fails with ``Assay.get_url()`` (#2559, #2565)
     - Vue3app study tab inactive after study or assay URL init (#2561)
+    - Vue3app initialization fails with legacy assay URLs (#2563)
 
 
 v1.3.2 (2026-09-17)